zsh: Add ability to ignore patterns to zcompile
Added ZCOMPILE_IGNORE_PATTERNS array that modules can add patterns to
in order to declare that the matched files/directories should not be
zcompiled.

This fixes a bug causing a failed compile on one of z-sy-h's test files,
as it was defining 'ls()', which conflicted with our ls alias. Making it
'function ls()' would have fixed the failed compilation, but these files
didn't need to be compiled in the first place.