vim: Update Neovim Python provider settings
Taking the advice of `:h python-virtualenv`, a new data directory has
been added called 'pynvim-venvs' which will hold a virtualenv for
Python2 and Python3 whose sole purpose is to hold an installation of
pynvim. The `python_host_prog` variables are set to the binaries in
these virtualenvs.

The existence of the 'pynvim-venvs' directory is ensured by a wrapper
similar to the other data directories. If the directory was created, for
now it will only warn that the virtualenvs should be created; eventually
this will be automated.