vim: Initial setup for native nvim LSP client
Finally decided to check out the nightly nvim builds and have since done
a deep dive for the past 3 days on the new Lua stuff, LSP client, and
reevaluating some plugins.

This is just a rather basic setup for now, relying on the default
handlers for the LSP requests. I intend to experiment with some of the
existing up-and-coming LSP-related plugins, as well as possibly
something of my own. Perhaps I'll dive further into vim plugin dev and
an excuse to learn Lua.