refactor(nvim): Use null-ls for formatting and linting
Now using null-ls to power our linting and formatting. This is somewhat
because the formatting plugin we were using changed focus to only
interact with LSP, but mostly because leveraging the LSP client for this
is better anyway.

null-ls is beautiful and has builtin support for tons of tools and
seamlessly integrates them into both Neovim's diagnostics and LSP
facilities. I get on-demand formatting, formatting on write,
diagnostics, code actions, and more, and melds with LSP configuration
and handlers in an expected way.