refactor(nvim): Improve logging utility functions
`get_module_name` now takes an optional `level` argument that is passed
to `debug.getinfo`. It now returns nil when an appropriate module name
cannot be determined.

`util.log` is now private and schedules the `nvim_echo` call to be safe
in the event loop. It also omits the module name if `get_module_name`
returns nil.

`util.error` will print a traceback in addition to an error message.