feat(zsh): Begin using zcomet
My homebrew zsh config framework has served me well enough for my needs
for a long time, but I've decided to finally use a plugin manager for
a few reasons:

- I want to use more external plugins, and keeping them as submodules is
  lame for the same reason that keeping Vim plugins as submodules is
  lame
- Better framework techniques have been developed and techniques once
  thought beneficial have been found to be problematic (bless romkatv)
- I don't particularly want to maintain plugin management code on top of
  my shell configuration anymore; others can do that better than I can.

I have decided on zcomet for now, with antidote a close second. The
former is nearly identical in performance to the latter while being
a bit slimmer.