refactor(nvim): Improve LSP buffer autocmds
LSP buffer autocmds are now defined in augroups that are specific to
both the buffer and LSP client that caused their creation in the first
place. This makes it easier to namespace them and ensure that no
duplicates are created. Thanks to mfusseneggar for this technique.

Also reworked my automatic signature help to trigger on
a `TextChangedPre` autocmd instead of mapping the trigger characters
directly. This was most notably trampling over the autopairs mapping for
`(`.