refactor(nvim): Overhaul lua_ls configuration
The baseline configuration for LLS should be minimal and not assume a
Neovim environment. Loading all of my plugins' lua files into the
workspace library was entirely unnecessary and caused a ton of
"latency". This was mostly due to misunderstandings and half-efforts on
my part at the time of initial setup.

We now leverage lspconfig's `on_new_config` callback to only load in the
Neovim specific stuff when we're actually editing something from our
config (fancy that). Additionally, we now ensure that the `root_dir` is
always set to the nvim config dir instead of sometimes being rooted at
the repo root depending on the file edited.