feat: Add Ansible setup
It was about time I stopped putting off setting up some configuration
management, so I've finally done so with Ansible. Initially the idea was
just to deploy my dotfile links for the most part, but it grew into
a full personalized environment setup.

To start with, I have two playbooks:
- `bootstrap.yml`: Does everything, including `nvim_bootstrap.yml`
- `nvim_bootstrap.yml`: Installs and bootstraps my Neovim setup

The do-everything bootstrap playbook will take me from a fresh user
directory and perform all of the usual mundane tasks of setting up my
personal environment, configurations, and tools. A high level overview
of what is done is as follows:

- Home directory setup
  - Ensures basic directory structure like XDG dirs, Projects root, src
    dir for arbitrary repos and building, etc.
  - Ensure proper permissions on .ssh
  - Purge unnecessary bash-related files
  - Deploy and ensure dotfile links
- Software management
  - Manage repositories depending on distro
  - Install various utilities
  - Sparse-checkout the Toyko Night repository to use the theme in other
    utilities
  - Ensure at least some version of Python 3 is installed
  - Bootstrap Python venv in .local/opt for user installs
  - Purge unwanted packages
  - Install desktop-oriented applications if running in a graphical
    environment

With Ansible, a lot of the bootstrap things I was doing in
`bootstrap.lua` as part of my config can be done much more robustly with
Ansible. As a result, much of that logic in Lua was moved into the new
`nvim` Ansible role and is run by the `nvim_bootstrap` playbook, which
does the following:

- Install build dependencies
- Clone and checkout the nightly tag
- Build and install
- Ensure various directories exist
- Download lazy.nvim and sync plugins
- Set up providers