feat(nvim): Switch from null-ls to nvim-lint + conform.nvim
Resolves #35. In the wake of null-ls being archived, I might as well be
proactive and switch linting/formatting solutions now. Farewell,
null-ls, you were incredible.

As a replacement, I've opted for nvim-lint paired with conform.nvim;
reasoning is outlined in #35. As a consequence of this change, some
related things were restructured. The `format` module was moved out of
the `plugins.lsp` namespace and into the more general `zeroknight`
namespace, as well as some tweaks that allow plugins to hijack it.
Further, since formatting is no longer "just" an LSP thing, the mappings
for formatting were updated. A couple new utility functions were added
as well.