refactor(nvim): Switch to jls for Python
Here we go again, trying a different Python language server, this time
the Jedi LS. It seems more stable all around and more performant than
pylsp, and doesn't have the extra weight of trying to integrate
linters and formatters into it when I can use them through other means.
As far as implementation of LSP features goes, it seems very on-par
with pylsp save for auto-import... not that it really worked in pylsp
anyway.