feat(winterm): Add initial Windows Terminal configuration
Utilizes the "Fragments" system to break up and self-document bits of Windows
Terminal configuration. The initial state of settings.json is a bit
indeterminate and broken as-is. It is intended to be *just* the changes from
default that I want and nothing more. Too lazy to manage it right now, though.