feat(nvim): Set 'winborder' option
This option sets the default border style for floating windows. I didn't
know of this option until now, so going forward, I intend to rely on it
instead of explicitly setting it in every applicable plugin.

Of course, not every plugin may know of/respect this setting, so it will
likely still be useful to explicitly set my preference on an as-needed
basis.