feat(nvim): Add powershell.nvim
This plugin manages many of the idiosyncrasies of PSES to provide
a fuller LSP/DAP experience for PowerShell in Neovim.

Starting and stopping the server needs to managed by powershell.nvim
since it handles some session state and ensuring the PS extension
terminal is started. In lieu of this, my custom `disabled` keys in my
LSP server configurations is now used as the argument to
`vim.lsp.enable()` as opposed to whether it's configured at all.

Now any servers I care to keep configuration for will always be
configured as such as far as the LSP client is concerned, but
whether/when they are started can now be controlled more flexibly.