refactor(ansible): complete refactor of the software role
As my experience with Ansible grew, the more apparent it became that
`software` was far too generic to be a useful role, even after the
recent restructuring. It has now been fully deconstructed into a more
sensible set of roles.

Most tasks ended up being re-classified as one of two higher-level
roles: `shell` or `desktop`, whose responsibilities are self-evident.
Most package installations are done in a single variable-driven
`package` install task. Further configuration for different utilities is
performed in a separate role.

This feels to me to be far easier to extend and maintain going into the
future. At least it does right now, anyway...

Resolves #57.