fix(ansible): dotfiles linking strategy
With some exceptions, only leaf nodes (files) under dotfiles are
deployed as symlinks. The old approach of lazily linking entire
directories as the rule had some issues:

- The programs could create additional files in the linked directories,
  consequently making the working tree in the repo dirty
- Directories that existed in home prior to being added to the repo
  would fail to deploy via Ansible later, as it refuses to replace
  existing files

While the logic for deploying the symlinks became a bit more
complicated, it resolves the issues above and *should* be seamless and
idempotent going forward, even when new configs are added to the repo.

Resolves #62.